Armed with LoRA to classify the digits for NN
Train the network only for 1 epoch to simulate a complete general pre-training on the data
Since the network performs poorly on the digit 9, freeze all the parameters of the original network and only fine tuning the ones introduced by LoRA. Then fine-tune the model on the digit 9 and only for 100 batches.
After fine tune, the model does a better job on classify digit 9 but get much worse on 3 and 7...
Pressing down on the gourd only makes the dipper pop up; it's impossible to take care of everything perfectly. However, if a specific task only cares about achieving a 9 (out of 10), then this kind of efficient fine-tuning is absolutely feasible.